+++ CHANGELOG.md
@@ -4,6 +4,23 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: Add `Field` component and sub-components:
+  - `Field.Label`
+  - `Field.Input`
+  - `Field.RadioButtonSet`
+  - `Field.Select`
+  - `Field.ErrorMessage`
+- Feature: Add some icons:
+  - `BancontactIcon`
+  - `GiropayIcon`
+  - `IdealIcon`
+  - `MaestroIcon`
+  - `PayPalIcon`
+  - `SofortIcon`
+- Feature: Add `k-TextInput--sixDigits` class.
+- Fix: Update `k-inputSet` vertical spacing.
+- Fix: Update `k-RadioButtonSet` line-height.
+
 ## [1.0.0] - 2018-12-11
 
 - First release!